# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
    "/v1": {
      "get": {
        "operationId": "AppController_getHello_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "App"
        ]
      }
    },
    "/v1/users": {
      "get": {
        "operationId": "UserController_getAllUsers_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get all users (protected)",
        "tags": [
          "Users"
        ]
      },
      "post": {
        "operationId": "UserController_createUser_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created successfully"
          }
        },
        "summary": "Create a new user",
        "tags": [
          "Users"
        ]
      }
    },
    "/v1/auth/login": {
      "post": {
        "operationId": "AuthController_login_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Login and get JWT token",
        "tags": [
          "Auth"
        ]
      }
    }
  },
  "info": {
    "title": "Swagger Demo API",
    "description": "NestJS Swagger learning project",
    "version": "1.0",
    "contact": {}
  },
  "tags": [],
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Akanksha",
            "description": "Full name of the user"
          },
          "email": {
            "type": "string",
            "example": "akanksha@gmail.com",
            "description": "Email address"
          },
          "password": {
            "type": "string",
            "example": "password123",
            "minLength": 6
          },
          "age": {
            "type": "number",
            "example": 25
          }
        },
        "required": [
          "name",
          "email",
          "password"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "example": "akanksha",
            "description": "Username"
          },
          "password": {
            "type": "string",
            "example": "123456",
            "description": "Password"
          }
        },
        "required": [
          "username",
          "password"
        ]
      }
    }
  }
}